refactor(cli): split _cli monolith into command modules (line-limit)
